Modified commands and added command descriptions

+++ src/DBHelper/orderDB.py
@@ -1,5 +1,5 @@
 import sqlite3
-from DBHelper import DATABASE_DIRECTORY, ORDER_COMMAND_TABLE
+from DBHelper import DATABASE_DIRECTORY, ORDER_COMMAND_TABLE, vacuum
 from DBHelper.usersDB import ID
 
 #Import attribute names
@@ -32,15 +32,33 @@ def get_initiative_order(user_id: int) -> list:
 
     return order
 
-def remove_one_order(user_id: int, char_name: str):
+def get_one_order(user_id: int, char_name: str):
     conn = sqlite3.Connection(DATABASE_DIRECTORY)
     cursor = conn.cursor()
 
-    cursor.execute(f"""DELETE FROM {ORDER_COMMAND_TABLE}
-                        WHERE {USER_ID} = {user_id} AND {NAME} = '{char_name}';""")
+    cursor.execute(f"""SELECT *
+                        FROM {ORDER_COMMAND_TABLE}
+                        WHERE {USER_ID} = {user_id} AND {NAME} = '{char_name}';
+                        """)
+    selection = cursor.fetchone()
 
-    conn.commit()
     conn.close()
+    return selection
+
+
+def remove_one_order(user_id: int, char_name: str) -> bool:
+    if get_one_order(user_id, char_name) != None:
+        conn = sqlite3.Connection(DATABASE_DIRECTORY)
+        cursor = conn.cursor()
+
+        cursor.execute(f"""DELETE FROM {ORDER_COMMAND_TABLE}
+                            WHERE {USER_ID} = {user_id} AND {NAME} = '{char_name}';""")
+
+        conn.commit()
+        conn.close()
+        return True
+    else:
+        return False
 
 def clear_user_order(user_id: int):
     conn = sqlite3.Connection(DATABASE_DIRECTORY)
